requirements.txt
streamlit>=1.35
pymilvus>=2.4.6
pypdf>=4.2.0
numpy>=1.26
python-dotenv>=1.0
reportlab>=4.2
openai>=1.40.0
sentence-transformers>=2.7.0
requests>=2.32